+++ lib/courses/home.dart
@@ -2,6 +2,7 @@ import 'dart:math';
 
 import 'package:dev_doctor/models/course.dart';
 import 'package:dev_doctor/models/server.dart';
+import 'package:dev_doctor/widgets/appbar.dart';
 import 'package:dev_doctor/widgets/image.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_modular/flutter_modular.dart';
@@ -115,7 +116,7 @@ class _CoursesPageState extends State<CoursesPage> {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-        appBar: AppBar(title: Text("courses.title").tr(), actions: [
+        appBar: MyAppBar(title: "courses.title".tr(), actions: [
           IconButton(
               icon: Icon(Icons.search_outlined),
               onPressed: () {