+++ lib/models/collection.dart
@@ -29,7 +29,7 @@ class BackendCollection {
         var current = _box.values.toList().indexOf(url);
         if (current != -1) index = _box.keyAt(current);
       } else if (url == null) url = _box.get(index);
-      data = await loadFile("$url/config");
+      data = await loadFile("$url/config") ?? {};
     } catch (e) {
       print(e);
     }