+++ lib/models/server.dart
@@ -68,16 +68,17 @@ class CoursesServer {
   Future<CoursesServer> toggle() => added ? remove() : add();
 
   static Future<CoursesServer> fetch({String url, int index, BackendEntry entry}) async {
-    var data = Map<String, dynamic>();
+    var data = <String, dynamic>{};
     try {
       if (index == null) {
         var current = _box.values.toList().indexOf(url);
         if (current != -1) index = _box.keyAt(current);
       } else if (url == null) url = Hive.box<String>('servers').get(index);
-      data = await loadFile("$url/config");
+      data = await loadFile("$url/config") ?? {};
     } catch (e) {
       print(e);
     }
+    data['courses'] = data['courses'] ?? [];
     data['entry'] = entry;
     data['url'] = url;
     data['index'] = index;